# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "RSS Transformation Service API",
      "version": "1.0.0",
      "description": "API for converting RSS feeds to other formats",
      "contact": {
        "name": "Utviklere i NB-TL",
        "email": "TL-utviklere@nb.no"
      }
    },
    "externalDocs": {
      "description": "NLB API documentation",
      "url": "https://api.nlb.no/documentation"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "tags": [
      {
        "name": "Health",
        "description": "Health check"
      },
      {
        "name": "RSS",
        "description": "Convert RSS feeds"
      }
    ],
    "paths": {
      "/health": {
        "get": {
          "tags": [
            "Health"
          ],
          "summary": "Health check",
          "description": "Check if the service is up and running",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/convert/{outputFormat}": {
        "post": {
          "tags": [
            "RSS"
          ],
          "summary": "Convert RSS feed to given format",
          "description": "Convert RSS feed to given format",
          "parameters": [{
            "in": "path",
            "name": "outputFormat",
            "description": "Format to convert to",
            "required": true,
            "type": "string",
            "enum": [
              "dtbook",
              "xhtml"
            ],
            "default": "dtbook"
          }],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "feedUrl": {
                      "type": "string",
                      "format": "url",
                      "description": "URL to RSS feed"
                    },
                    "metadata": {
                      "$ref": "#/definitions/Metadata"
                    },
                    "options": {
                      "$ref": "#/definitions/Options"
                    }
                  },
                  "required": [
                    "feedUrl",
                    "editionId"
                  ]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/xml": {
                  "schema": {
                    "type": "string",
                    "description": "NLBPUB XML"
                  }
                }
              }
            },
            "400": {
              "description": "Bad request"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
    },
    "definitions": {
      "Metadata": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Edition ID"
          },
          "language": {
            "type": "string",
            "description": "Language code"
          },
          "title": {
            "type": "string",
            "description": "Title of the publication"
          },
          "description": {
            "type": "string",
            "description": "Description of the publication"
          },
          "responsibilityStatement": {
            "type": "string",
            "description": "Responsibility statement"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Genre"
            }
          },
          "audiences": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Audience"
            }
          },
          "available": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the publication is available"
          },
          "publishedYear": {
            "type": "integer",
            "description": "Year of publication"
          },
          "registered": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the publication was registered"
          },
          "pages": {
            "type": "integer",
            "description": "Number of pages"
          },
          "volumes": {
            "type": "integer",
            "description": "Number of volumes"
          },
          "publishedLocation": {
            "type": "string",
            "description": "Location of publication"
          },
          "publisher": {
            "type": "string",
            "description": "Publisher"
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Subject"
            }
          },
          "subTitle": {
            "type": "string",
            "description": "Subtitle"
          },
          "edition": {
            "type": "string",
            "description": "Edition number"
          },
          "isbn": {
            "type": "string",
            "description": "ISBN"
          }
        }
      },
      "Genre": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name"
          },
          "identifier": {
            "type": "string",
            "description": "Identifier"
          }
        }
      },
      "Audience": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Identifier"
          },
          "age_from": {
            "type": "integer",
            "description": "Age from"
          },
          "age_to": {
            "type": "integer",
            "description": "Age to"
          },
          "name": {
            "type": "string",
            "description": "Name"
          }
        }
      },
      "Subject": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Identifier"
          },
          "name": {
            "type": "string",
            "description": "Name"
          },
          "location": {
            "type": "string",
            "description": "Location"
          },
          "dewey": {
            "type": "string",
            "description": "Dewey"
          }
        }
      },
      "Options": {
        "type": "object",
        "optional": true,
        "properties": {
          "includeAbout": {
            "type": "boolean",
            "optional": true,
            "description": "Include about page"
          },
          "includeToc": {
            "type": "boolean",
            "optional": true,
            "description": "Include table of contents",
            "default": false
          }
        }
      }
    }
  }